requirements.txt
alembic==0.8.4
bleach==1.4.2
cffi==1.4.1
ckanapi==3.5
colorama==0.3.5
countrycode==0.2
coverage==3.7.1
cryptography==1.1.2
docopt==0.6.2
html5lib==0.9999999
idna==2.0
Mako==1.0.3
MarkupSafe==0.23
ndg-httpsclient==0.4.0
nose==1.3.0
pyasn1==0.1.9
pycparser==2.14
pyOpenSSL==0.15.1
python-editor==0.5
python-slugify==1.1.4
python-termstyle==0.1.10
rednose==0.4.3
requests==2.8.1
scraperwiki==0.5.1
simplejson==3.8.1
six==1.10.0
SQLAlchemy==1.0.11
termcolor==1.1.0
Unidecode==0.4.18
